Added caching for components#index.json, closes #164
