+++ app/workers/reindex.rb
@@ -5,5 +5,10 @@ class Reindex
 
   def perform(force = false)
     Build::Converter.index!(force)
+
+    # Overwrite the components_json cache key
+    Rails.cache.write('components_json') do
+      ComponentHelper.generate_component_json
+    end
   end
 end
